mv test_*.py to test_math_ops.py and fix bugs
